// x-pack/platform/plugins/shared/cases/public/components/case_form_fields/utils.d.ts
export declare const validateEmptyTags: ({ value, message, }: {
    value: string | string[];
    message: string;
}) => {
    message: string;
} | undefined;
export declare const validateMaxLength: ({ value, message, limit, }: {
    value: string | string[];
    message: string;
    limit: number;
}) => {
    message: string;
} | undefined;
export declare const validateMaxTagsLength: ({ value, message, limit, }: {
    value: string | string[];
    message: string;
    limit: number;
}) => {
    message: string;
} | undefined;